cleaned up empty files and deprecated code in dataAnalysis folder; kept utils.py as it's referenced by other files
